refactor: add useAutoSelectFirstEnvironment hook to remove duplicates
